Add exported flag to receiver in FastPairManager
Android T allows apps to declare a runtime receiver as not exported
by invoking registerReceiver with a new RECEIVER_NOT_EXPORTED flag;
receivers registered with this flag will only receive broadcasts from
the platform and the app itself. However to ensure developers can
properly protect their receivers, all apps targeting U or later
registering a receiver for non-system broadcasts must specify either
the exported or not exported flag when invoking #registerReceiver;
if one of these flags is not provided, the platform will throw a
SecurityException. This commit updates the exposed receiver in
FastPairManager with RECEIVER_EXPORTED since it is registering
for a bluetooth broadcast which is sent by a UID other than
the current app or system UIDs; this also allows the non-system
fast pair actions to be sent from outside the local app.

Bug: 234659204
Test: atest FastPairManagerTest
Ignore-AOSP-First: nearby_not_in_aosp_yet
Change-Id: Id6759437cb3647bb0853ad0c3e6f59d32cfdaaba

Former-commit-id: 02362557f8f1062f2bce7daab29b13207b4ff901