Use a longer timeout for service registration
NsdManager may take some time to register a service when the device
hostname is already used on the network, as after probing it needs to
choose another hostname and try again. As the default hostname is
"Android" and the alternatives are "Android-2", "Android-3", ..., this
may take several seconds: RFC6762 8.1 says probing should be 3 packets
with a 250ms interval.

Relax the timeout to 10s, which should allow ~10 conflicts before the
test starts failing; at that point, MDNS isn't reliably testable on the
network.

Bug: 248121960
Bug: 257459913
Bug: 257393724
Test: atest
Change-Id: I31e47787ed9661524edd6b676342fb9a442408a8

Former-commit-id: 22d349083ce71cebf2d571527324691c65951fc4