cronet: delete cronet prebuilts
Cronet is now built from source, so this code is not used anymore.

This CL is a cherry pick of aosp/2388908 to master
aosp/2388908 was not merged to the master and downstream branches since
ag/21052071 was merged to master and downstream branches with empty
changes by [automerger skipped] CLs.

Test: TH
Ignore-AOSP-First: Cherry-pick of aosp/2388908
Change-Id: I28aa8499583f51549de858ac89c7e08f9d1db0de

Former-commit-id: c8680fdcc3ad2fb91730e6d12a7f68cbe95657bb