Move mdns package to the right directory
All the files in the directory are in the
com.android.server.connectivity.mdns package, but they were put in the
com/android/server/mdns directory (no "connectivity").

This is a no-op as the directory is not used at build time, but fixing
the inconsistency avoids problems in IDEs among others.

Test: m
Change-Id: I7acb6d7d5ea0bd332e8f9503b7e4a976246e355d

Former-commit-id: 26774e5206ccfb3549bfb38999ead09dc0839475