Use current shims in service-connectivity
Instead of NetworkStackApiStableShims, use NetworkStackApiCurrentShims
as this is a development branch.

Use java_defaults in the "merge conflicts expected" section, so that
module release branches can use the same defaults, but set it to
NetworkStackApiStableShims.

The shims are removed from FrameworksNetIntegrationTests, as they are
already available there through service-connectivity-pre-jarjar. Keeping
them as a dependency would make FrameworksNetIntegrationTests use both
Stable shims (directly) and Current shims
(through service-connectivity-pre-jarjar).

Ignore-AOSP-First: this needs to be submitted separately in each branch,
                   with downstream branches first as there are merge
		   conflicts in Tethering/apex/Android.bp that cannot be
		   automerged.

Bug: 266205506
Test: m
Change-Id: I88228152834a1c06830fb51e868fb8e3d8c47519

Former-commit-id: 921290b49d109360aae5dc39effaa3b0e691f65a