Log when a TCP connection keeps up an auto keepalive
This shows the owner UID, the destination and the socket
cookie. The socket cookie can then be looked up in
dumpsys netstats to see what is the tag.

This is only logged if log level is at least debug for the
AutomaticOnOffKeepaliveTracker tag :
adb shell setprop log.tag.AutomaticOnOffKeepaliveTracker DEBUG

Test: manual
Change-Id: I54f3a99ab1483d50ee71cb158a54f67fe640b3fc

Former-commit-id: 9b1820103556df5eaaea89d1b3b6e565cd6709b9