gn2bp: Duplicate modules reachable from test targets
* The current testing situation for Cronet is dire due to the presence of Cronet on the platform,
As cronet is on the platform, we can't run any unit-tests as they'll link against the platform version
and not the statically-linked cronet. We could avoid this by jarjaring the code however this is impossible.

* It's impossible to jar-jar the code as the C++ code which does JNI expects jar-jared java classes. So the
only possible solution here would be to duplicate the C++ modules, one for testing(with no jarjar) and another
for platform(with jarjar)

* Java sources/actions are not collected for testing modules until a proper solution has been figured out.

Test: None
Change-Id: I20bbad13d83e9e3c7956f028f3f0493b24d0f8bc

Former-commit-id: edb2fd014bb6d6fdb7d3131cb267e788c24bd0a6