cronet import: use public chromium repository as a source
rpc://chromium/chromium/src contains a partial mirror of the Chromium
repo. Unfortunately, the mirror cannot be used as it seems to have been
created with a different gclient config (i.e. target_os is not set to
'android').

Additionally, do not use partial_fetch as gclient requires a full
checkout.

Test: run import script.
Change-Id: Iac185d3af38d95ec394551f3ca1487e65cf6c975

Former-commit-id: 3cefd0b8d1bd2716ac42e8b23909c962a2724b17