Update NsdManagerTest for newer backend usage
The new backend requires LinkProperties that reflect the actual
addresses on the interface. Make sure the test sets the addresses on the
test NetworkAgents that are used.

Also the new backend properly differentiates between boolean attributes
that have no value, and attributes with an empty value. As the new
backend is currently being enabled on U+, verify the new behavior in
that case.

Bug: 272197605
Test: atest NsdManagerTest
Change-Id: I8540af27bdd84aff84df9faba7327a8e9af802ef

Former-commit-id: d89289b31f89c1d40344f26a09008f955236e1fa