Verify Kernel MOBIKE with V6 UDP encapsulation
This commit adds tests to verify that SAs with UDP encapsulation
can be updated from current IPv6 addresses to new IPv6 addresses.

This commit also improves CTS to verify that SA update can work
properly when the UDP encapsulation socket is rebinded to a
new network.

Verified on kernel 5.10.149 and 5.10.104

Bug: 266149275
Test: atest IpSecManagerTunnelTest(new tests added)
Change-Id: I9b5c961808c2f368911192003974a95643aa5181

Former-commit-id: f0cc850de85ccb9fac28355fd1db6f3853916c1f