Use consistent addresses in SSLCertificateSocketFactoryTest.
Prior to this change, the expected local and remote
InetAddresses for connecting to the test host were determined
differently, the peer address arbitraily chosen from
InetAddress.getAllByName(), but the local address determined
by actually makeing a test connection with a plain socket.

This could potentially lead to one being an IPv4 address and
the other an IPv6 one, leading to connection failures.

Bug: 254385400
Test: atest CtsNetTestCasesLatestSdk:android.net.cts.SSLCertificateSocketFactoryTest
Change-Id: Ib3ec5d4881a5db2c5141e2dd9d6053f698d1ca85

Former-commit-id: 6ed4976ae5481b795950d5b1e275e6d3a3463956