Set dialog background activity start mode on U+
Android U requires pending intent background start mode to be specified,
when using PendingIntent.send. Set it for the "notify as dialog"
feature which uses PendingIntent.send.

PendingIntent.send() just calls the 7 argument version used here with
null/0 in AOSP, so if this is kept unmodified on the actual device,
this is a no-op on T-.

Bug: 270037904
Test: atest
Change-Id: I64ae0214d87224ad61bc776755d4b6792763b875

Former-commit-id: 985c3c73900d9f121fc61a2aaac21b17cbc7048d