Stop verifying wifi disconnect on T+ devices
Wifi remembers the wifi enabled state under airplane mode in the
Settings.Secure.WIFI_APM_STATE setting for the airplane mode
enhancement. Wifi will not be disconnected if user enables
the wifi under airplane mode. This behavior breaks the
assumption in testSetAirplaneMode. The test expects to always
receive the WiFi network disconnect after turning airplane
mode on.

The wifi enhancement was done with hidden method which should not
be referred in the cts. The intention of testSetAirplaneMode was
to verify the permission required for setting ariplane mode.
This change should be no harm for original intention.

Bug: 278150162
Test: atest CtsNetTestCases
Change-Id: I3458679378f0c12c26051f3dac7fcf155c1f469c

Former-commit-id: c346fc8a0d8330aa7cf033a8554233a90e683164