Increase socket timeout to 10s
Some tests have been observed to run at a time where validation
probes take 1~2s to complete. A 2s socket timeout is too short in
such cases.

Use 10s instead, as this matches the timeout used for validation in
general. This standardizes the timing expectations when connecting
to connectivitycheck.gstatic.com, both for validation and for this test.

Bug: 284533985
Change-Id: I36cf9cc05cec15bf66b8cfd450e3b9955f487417
Test: atest

Former-commit-id: 1191ee2df9414ff9256166dd3744b3df134aceab