[Cronet] Cancel stream before shutdown
This fixes the flakiness. Because we call to a real server,
the server may not be reachable and our assumption would fail.
When this happens, the stream hangs and engine shutdown fails with
active request error. Hence cancel the stream before calling
shutdown.

Also @Ignored this test despite the fix as a safety measure to meet
the SLO. Will re-enable afterwards.

Bug: b/292298108
Test: atest NetHttpCoverageTests --test-filter
BidirectionalStreamTest#testBidirectionalStream_GetStream_CompletesSuccessfully
Change-Id: Iaae61fc62a148b58af86fb29c248399041785f9b

Change-Id: Iea9ef0b8468ec6fa05c306ea394485cec706b9ef

Former-commit-id: 54db2af362de44ff5bd4cc075910afdaf94b70fe