Add sim card requirement for net permission tests
The network permission tests perform a connectivity check before
running, which means that a SIM card is required to check for
mobile data. However, there is no SIM card token sharding defined
in the AndroidTest.xml to make sure that the SIM card is
available for testing. Therefore, add the definition to the
AndroidTest.xml file.

Bug: 295708437
Bug: 295611982
Test: m
Change-Id: I20201ea8c7272771fdc59752aa2aae22b14d1e34

Former-commit-id: 7615075d12266dda159a991d9fabd42c3a0b15f5