Skip tests that require platform fix before S
In T, there are changes to fix threading issues in the platform
codes. Those tests that verify the behavior on starting and
stopping the VpnProfile will have chances to hit the issue.
For devices running with S or before S, there is no way to fix
it until upgrading to T+ because the fixes were landed in the
platform code. These tests should only need to run with the
latest sdk theoretically because tests verify how platform
code behave.

Add annotation on the tests as other testStartStopVpnProfile*
tests to at least run on the platform that is more reliable.

Bug: 297703895
Test: atest Ikev2VpnTest
Change-Id: I3adb1f7a24ddbbbdbc4f5d7f0ea66502f3c93044

Former-commit-id: 98a505a253a40a19d4e6d90bf46a2ac36a283d7f