Assign default bug component to targets in this directory.
This CL is being created to improve the test attribution in android. It has been 
found that the tests owned by this OWNERS are not not being attributed to a
buganiser component.
This is part of a bigger effort to attribute all of the aosp code. 

If you think that the buganiser component is not correct please update
this CL with correct component and drop a +2.
We will use the updated component and merge the CL

Bug: 309090038
Change-Id: Ie6839c3f6d07a554c60d544acea712d8de76b6b2
Test: N/A

Former-commit-id: 5e157ce460d49b58fc10c5df2fc727f8c8c8a568