Specify aconfig_declaration modules in java_sdk_library
java_sdk_library modules should specify the aconfig_declarations modules
in order to generate the exportable stubs (i.e. the stubs that exclude
the disabled aconfig flags) by providing the information about the
aconfig flag values.

The exportable stubs will be used when generating the module sdk
snapshots and to generate the hiddenapi flags, in order to prevent
exposing the disabled apis in runtime.

Test: m framework-connectivity-t
Bug: 317290555
Change-Id: Ieb8f41921054ed6f62db2edfc7d695ba0a2a0055

Former-commit-id: 60b21f877f88a6848ee6d0f0e5316362714b9d25