Implement offload priority
Implement the TODO to assign priority to services based on an
overlayable configuration.

This changes the default priority from 0 to MAX_VALUE, as higher
priority values are less prioritized, so services not in the priority
list should have a high value. There is currently no user of the API
depending on the priority, so the change should not affect clients.

Test: atest
Change-Id: I3aae9b5fa7756de64c63974c192e4c96b632ce4e

Former-commit-id: 5c9d6cbf163325693e6ff4151472f39cbcca3a25