[Thread] discover and resolve the meshcop service only on test network
The test case is flaky because the service is unexpectedly discovered
after it's already unregistered. Probably this is due to that on the
WiFi interface NsdManager somehow didn't hear the goodbye PTR for the
service so it can still 'discover' the service. Limiting the discovery
on test network may help avoid the flake because test network is
cleaner.

Bug: 331878313
Test: https://android-build.corp.google.com/builds/abtd/run/L17400030002839695

Change-Id: Icc09aae00cd90f64428387b97dbd9743140641e5

Former-commit-id: a449374772ef9e8f2ee701fd5bea4f2f2069224d