Update NetworkRequestTest#testDefaultCapabilities
testDefaultCapabilitieswas was added when aosp/2539630 introduced
DEFAULT_FORBIDDEN_CAPABILITIES to avoid existing users getting callbacks
from local networks.
But aosp/2539630 was reverted and aosp/2936050 fixed this issue by
introducing compat flag.
testDefaultCapabilities was missed to be cleaned up due to the
annotation.
This CL updates testDefaultCapabilities to verify the default
capabilities and default forbidden capabilities are not changed on U-.

Test: NetworkRequestTest
Bug: 332466530
Change-Id: I62c990d0f2c73d159e7c93827bbec14d4d0c0f45

Former-commit-id: 1cf04c96dadff7a74ff50cca36ec46b16831148e